Created abstract RegularCommand class.
